Block signals with no valid TP target + fix float years in sync endpoint
analysis.py:
  - Add no-target gate in generate_signal(): when risk_context returns
    tp_source='no_3r_target' (no valid TP found), skip the signal entirely
    instead of entering with TP=entry (0 reward, full risk — guaranteed drain)
  - This was the #1 identified drag from the backtest replay:
    ~28 BTC trades and ~30 ETH trades were no_3r_target entries
  - Live bot now only fires when a real structural TP level exists

app.py (/api/backtest/sync):
  - Change years param from int() to float() so 0.5, 1.5 etc. work
  - Enables 15m and 1h backtests to use shorter windows without the
    'not enough historical data' error that int(0.5)=0 was causing

+++ app.py
@@ -1866,7 +1866,7 @@ def backtest_sync():
     if not symbol.endswith("USDT"):
         symbol += "USDT"
     interval = body.get("interval", "4h")
-    years    = int(body.get("years", 2))
+    years    = float(body.get("years", 2))   # float supports 0.5, 1.5 etc.
 
     if interval not in VALID_INTERVALS:
         return jsonify({"error": f"Invalid interval. Use: {VALID_INTERVALS}"}), 400
